+++ src/error/wrap.rs
@@ -1,10 +1,10 @@
-use std::{error::Error as StdError, fmt};
+use std::fmt::{self, Display};
 
-use super::{Error, HelpMsg};
+use super::Error;
 
 pub trait ErrorWrap<T, E>
 where
-    E: StdError + Send + Sync + 'static,
+    E: Send + Sync + 'static,
 {
     /// Wrap an error value with additional context that is evaluated lazily
     /// only once an error does occur.
@@ -28,17 +28,14 @@ where
 
 impl<T, E> ErrorWrap<T, E> for Result<T, E>
 where
-    E: StdError + Send + Sync + 'static,
+    E: ext::StdError + Send + Sync + 'static,
 {
     fn wrap<C, F>(self, f: F) -> Result<T, Error>
     where
         C: fmt::Display + Send + Sync + 'static,
         F: FnOnce() -> C,
     {
-        self.map_err(|err| Error {
-            inner: anyhow::Error::from(err).context(f()),
-            help: None,
-        })
+        self.map_err(|err| err.ext_context(f()))
     }
 
     fn wrap_help<C, F>(self, f: F, help: &'static str) -> Result<T, Error>
@@ -46,10 +43,7 @@ where
         C: fmt::Display + Send + Sync + 'static,
         F: FnOnce() -> C,
     {
-        self.map_err(|err| Error {
-            inner: anyhow::Error::from(err).context(f()),
-            help: Some(HelpMsg::Static(help)),
-        })
+        self.map_err(|err| err.ext_context_help(f(), help))
     }
 
     fn wrap_help_owned<C, F>(self, f: F, help: String) -> Result<T, Error>
@@ -57,9 +51,94 @@ where
         C: fmt::Display + Send + Sync + 'static,
         F: FnOnce() -> C,
     {
-        self.map_err(|err| Error {
-            inner: anyhow::Error::from(err).context(f()),
-            help: Some(HelpMsg::Owned(help)),
-        })
+        self.map_err(|err| err.ext_context_help_owned(f(), help))
+    }
+}
+
+mod ext {
+    use crate::error::HelpMsg;
+
+    use super::*;
+
+    pub trait StdError {
+        fn ext_context<C>(self, context: C) -> Error
+        where
+            C: Display + Send + Sync + 'static;
+
+        fn ext_context_help<C>(self, context: C, help: &'static str) -> Error
+        where
+            C: Display + Send + Sync + 'static;
+
+        fn ext_context_help_owned<C>(self, context: C, help: String) -> Error
+        where
+            C: Display + Send + Sync + 'static;
+    }
+
+    impl<E> StdError for E
+    where
+        E: std::error::Error + Send + Sync + 'static,
+    {
+        fn ext_context<C>(self, context: C) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            Error::from(self).wrap(context)
+        }
+
+        fn ext_context_help<C>(self, context: C, help: &'static str) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            let mut err = Error::from(self).wrap(context);
+            err.set_help(help);
+            err
+        }
+
+        fn ext_context_help_owned<C>(self, context: C, help: String) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            let mut err = Error::from(self).wrap(context);
+            err.set_help_owned(help);
+            err
+        }
+    }
+
+    impl StdError for Error {
+        fn ext_context<C>(self, context: C) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            self.wrap(context)
+        }
+
+        fn ext_context_help<C>(self, context: C, help: &'static str) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            let mut err = self.wrap(context);
+            match err.help {
+                Some(HelpMsg::Owned(ref mut msg)) => {
+                    msg.push('\n');
+                    msg.push_str(help);
+                }
+                Some(HelpMsg::Static(msg)) => err.set_help_owned(format!("{}\n{}", msg, help)),
+
+                None => err.set_help(help),
+            }
+            err
+        }
+
+        fn ext_context_help_owned<C>(self, context: C, help: String) -> Error
+        where
+            C: Display + Send + Sync + 'static,
+        {
+            let mut err = self.wrap(context);
+            match err.help() {
+                Some(msg) => err.set_help_owned(format!("{}\n{}", msg, help)),
+                None => err.set_help_owned(help),
+            }
+            err
+        }
     }
 }