+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Multiple `Error` help messages.
+- Wrapping a `Result` with an `Error`.
+
 ## [0.1.1] - 2022-09-15
 
 ### Added